Rethink solver and integrator:
There are two things to be solved: The riddle of calculating a mechanical systems next state and within this riddle-solving-process, there is usually a system of equations to be inverted/solved in some way. The latter could be called "solve an function numerically".

Therefore, we should refer to two types of solves: SystemSolvers and FunctionSolvers.

In order to fix the solver class, I had to fix the integrators first. An integrator should define rules to calculate residuum and tangent. That's it. It should be simple. We did have redundancy in terms of the two integrator base classes... this redundancy is removed. However, the specific implementation of "get_residuum" and "get_tangent" is just copy-pasting of old function "calc_residuum_tangent". So the new function have to be refactored themselves.